backend/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
pydantic==2.5.0

# ML dependencies for training and inference
# Note: If you get DLL errors on Windows, try CPU-only version:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# For production deployment
# gunicorn==21.2.0
